+++ client/lib/js/server-features.js
@@ -37,23 +37,6 @@ var ABeamer;
             },
         },
     };
-    // @TODO: Use feature detection to determine which properties requires prefixes.
-    var BROWSERS = {
-        chrome: {
-            vendor: '-webkit-',
-            prefix: ['background-clip'],
-        },
-        firefox: {
-            vendor: '-moz-',
-            prefix: ['background-clip'],
-        },
-        ie: {
-            vendor: '-ms-',
-            prefix: ['background-clip'],
-        },
-    };
-    var curBrowserPrefix = BROWSERS.chrome.prefix;
-    var curVendor = BROWSERS.chrome.vendor;
     /**
      * Adds server property mapping and returns the server features.
      * Used by `Story.Create` method to assign the server property mapping.
@@ -64,54 +47,5 @@ var ABeamer;
         return serverInfo.features;
     }
     ABeamer._setServer = _setServer;
-    /**
-     * Maps an input CSS property into the current CSS property, adding a prefixed
-     * CSS property if necessary.
-     */
-    function _propNameToVendorProps(propName) {
-        propName = propName.replace(/(?:-webkit-|-moz-|-ie-)/, '');
-        if (curBrowserPrefix.indexOf(propName) !== -1) {
-            return [propName, curVendor + propName];
-        }
-        return [propName];
-    }
-    ABeamer._propNameToVendorProps = _propNameToVendorProps;
-    /**
-     * Initializes the `curBrowserPrefix` and `curVendor` based on the running browser.
-     */
-    function initBrowser() {
-        // this code is credit to https://stackoverflow.com/users/938089
-        // https://stackoverflow.com/questions/9847580/how-to-detect-safari-chrome-ie-firefox-and-opera-browser
-        // @ts-ignore
-        // Opera 8.0+
-        var isOpera = (!!window.opr && !!opr.addons) || !!window.opera || navigator.userAgent.indexOf(' OPR/') >= 0;
-        // @ts-ignore
-        // Firefox 1.0+
-        var isFirefox = typeof InstallTrigger !== 'undefined';
-        // @ts-ignore
-        // Safari 3.0+ "[object HTMLElementConstructor]"
-        var isSafari = /constructor/i.test(window.HTMLElement) ||
-            (function (p) { return p.toString() === "[object SafariRemoteNotification]"; })(
-            // @ts-ignore
-            !window['safari'] || (typeof safari !== 'undefined' && safari.pushNotification));
-        // @ts-ignore
-        // Internet Explorer 6-11
-        var isIE = /*@cc_on!@*/ false || !!document.documentMode;
-        // @ts-ignore
-        // Edge 20+
-        var isEdge = !isIE && !!window.StyleMedia;
-        // @ts-ignore
-        // Chrome 1+
-        var isChrome = !!window.chrome && !!window.chrome.webstore;
-        // added code
-        if (isFirefox) {
-            curBrowserPrefix = BROWSERS.firefox.prefix;
-            curVendor = BROWSERS.firefox.vendor;
-        }
-        else if (isIE) {
-            curBrowserPrefix = BROWSERS.ie.prefix;
-            curVendor = BROWSERS.ie.vendor;
-        }
-    }
 })(ABeamer || (ABeamer = {}));
 //# sourceMappingURL=server-features.js.map