+++ server/server-agent.js
@@ -113,8 +113,10 @@ var ServerAgent;
                 + sc.SERVER_SUFFIX + this.serverName + '&'
                 + sc.LOG_LEVEL_SUFFIX + this.logLevel + '&'
                 + sc.TELEPORT_SUFFIX + this.toTeleport.toString();
-            if (this.isVerbose) {
-                console.log("pageUrl: " + pageUrl);
+            // since this is an fundamental information it should always display
+            // even not on verbose mode.
+            if (this.logLevel !== 0) {
+                console.log("pageUrl: [" + pageUrl + "]");
             }
             return pageUrl;
         };